+++ src/app/directives/click-outside.directive.ts
@@ -1,0 +1,18 @@
+import { Directive, EventEmitter, ElementRef, HostListener, Output } from '@angular/core';
+
+@Directive({
+  selector: '[appClickOutside]'
+})
+export class ClickOutsideDirective {
+  @Output() clickOutside = new EventEmitter<MouseEvent>();
+
+  constructor(private eRef: ElementRef) { }
+
+  @HostListener('document:click', ['$event'])
+  clickout(event) {
+    if (this.eRef.nativeElement.contains(event.target)) {
+    } else {
+      this.clickOutside.emit(event);
+    }
+  }
+}